+++ run-tests.sh
@@ -1,0 +1,1 @@
+make check && make test-desktop && make build-android